Disable reinclude guard
For a while, I've been getting random errors that "variable lsa is
undefined", which made zero sense. It also doesn't appear all the time,
so I've been unable to track down what. Today, I found the source.

I initially assumed it was bad use of <SID> somehow, but this isn't
(fully) the case. Switching to <ScriptCmd> broke both command and
autocmd, which is bad. ScriptCmd was likely only intended for mappings,
and not commands and autocmds.

The error only appears when reloading vim. Removing the re-include guard
fixes it.

My assumption was that reloading is special with vim9, and somehow
results in the previously imported lsa being cleared. After writing this
part of the commit message, I went to check, and lo and behold, I'm
correct. SEe `:h vim9-reload`, which states:

> When loading a Vim9 script a second time all existing script-local functions
> and variables are deleted, thus you start with a clean slate.  This is useful
> if you are developing a plugin and want to try a new version.  If you renamed
> something you don't have to worry about the old name still hanging around.

This is super useful, and means the include guard is actually negative,
and prevents the plugin from being loaded after a reload. Fun, isn't
it?

TL;DR: include guards in vim9 are bad, as vim9 fixed the problem that
caused them to exist in the first place